requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
apptools==5.3.0
astroid==2.15.8
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
autoflake==2.1.1
black==23.3.0
blinker==1.8.2
build==1.2.2
CacheControl==0.14.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
comm==0.2.2
ConfigArgParse==1.7
configobj==5.0.9
contourpy==1.3.0
crashtest==0.4.1
cryptography==43.0.1
cycler==0.12.1
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.9
distlib==0.3.8
docker-pycreds==0.4.0
dulwich==0.21.7
envisage==7.0.3
exceptiongroup==1.2.2
executing==2.1.0
faiss-gpu==1.7.2
fast-pytorch-kmeans==0.2.0.1
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
freetype-py==2.5.1
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
-e git+https://github.com/jacintosuner/Grounded-SAM-2.git@336d924aae62a64440d5d205b4b54d482d87ab42#egg=groundingdino&subdirectory=../../../third_party/Grounded-SAM-2/grounding_dino
grpcio==1.66.2
h5py==3.12.1
huggingface-hub==0.25.1
hydra-core==1.3.2
identify==2.6.1
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
iniconfig==2.0.0
installer==0.7.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
isort==5.12.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keras==3.5.0
keyring==24.3.1
kiwisolver==1.4.7
lazy-object-proxy==1.10.0
lazy_loader==0.4
libclang==18.1.1
lightning==2.4.0
lightning-utilities==0.11.7
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mayavi==4.8.2
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy==1.3.0
mypy-extensions==1.0.0
namex==0.0.8
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvcc-cu12==12.3.107
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open3d==0.18.0
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.12.1
packaging==24.1
pandas==2.2.3
pandas-stubs==2.0.2.230605
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.11.1
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
portalocker==2.10.1
pre-commit==3.3.3
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.5
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycocotools==2.0.8
pycparser==2.22
pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git@2723c7fa4f2ead16ae1ce3d8afe977724bb8f87f
pyface==8.0.0
pyflakes==3.2.0
pyglet==2.0.17
Pygments==2.18.0
pylint==2.17.4
pynvml==11.5.3
PyOpenGL==3.1.0
pyparsing==3.1.4
pyproject_hooks==1.2.0
pyquaternion==0.9.9
pyrender==0.1.45
pytest==7.3.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytorch-metric-learning==2.6.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rich==13.9.1
-e git+https://github.com/jacintosuner/robot-grasp.git@1eaffba10f1cf1e41862335e4e6be2714de947c2#egg=robot_grasp
rpds-py==0.20.0
safetensors==0.4.5
-e git+https://github.com/jacintosuner/Grounded-SAM-2.git@336d924aae62a64440d5d205b4b54d482d87ab42#egg=SAM_2&subdirectory=../../../third_party/Grounded-SAM-2
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
SecretStorage==3.3.3
segment-anything==1.0
sentry-sdk==2.16.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
supervision==0.24.0
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==1.0.9
tokenizers==0.20.0
tomli==2.0.2
tomlkit==0.13.2
torch==2.4.1
torchaudio==2.4.1
torchmetrics==1.4.2
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
traits==6.4.3
traitsui==8.0.0
transformers==4.45.2
trimesh==4.4.9
triton==3.0.0
trove-classifiers==2024.9.12
types-pytz==2024.2.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.26.6
vtk==9.3.1
wandb==0.18.3
wcwidth==0.2.13
Werkzeug==3.0.4
widgetsnbextension==4.0.13
wrapt==1.16.0
yapf==0.40.2
yarl==1.14.0
zipp==3.20.2
